requirements.txt
# 核心依赖
torch>=2.0.0,<2.2.0
transformers>=4.30.0,<4.40.0
pydantic>=2.5.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
requests>=2.31.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pyarrow>=14.0.1
TA-Lib>=0.4.24

# 数据库
psycopg2-binary>=2.9.0
redis>=5.0.0
aioredis>=2.0.0  # 异步Redis支持
redis-py-cluster>=2.1.0  # Redis集群支持

# Web服务
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.5
websockets>=10.0
aiohttp>=3.9.0  # 异步HTTP客户端

# 安全
cryptography>=41.0.0
python-jose>=3.3.0
passlib>=1.7.0
bcrypt>=4.0.0  # 密码哈希

# 监控和日志
prometheus-client>=0.17.0
python-json-logger>=2.0.0
tensorboard>=2.15.0
wandb>=0.16.0
psutil>=5.9.0
py-spy>=0.3.0  # 性能分析
statsd>=4.0.0  # 指标收集
grafana-api>=1.0.3  # Grafana集成

# GPU优化
triton>=2.0.0
deepspeed>=0.10.0
bitsandbytes>=0.41.0
ninja>=1.11.0
apex>=0.1  # NVIDIA Apex优化
torch-tb-profiler>=0.4.0  # PyTorch性能分析

# 图像处理
Pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.0  # 图像增强

# 分布式训练
ray[train]>=2.7.0  # Ray分布式训练
ray[tune]>=2.7.0  # Ray超参数调优
horovod>=0.28.0  # 分布式训练
dask>=2023.0.0  # 分布式计算
dask-cuda>=23.12.0  # GPU支持

# 自适应学习
optuna>=3.4.0  # 超参数优化
bayesian-optimization>=1.4.0  # 贝叶斯优化
pytorch-lightning>=2.1.0  # 训练框架
torchmetrics>=1.2.0  # 训练指标

# 开发工具
pytest>=7.4.0
black>=23.12.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0

# 可视化
plotly>=5.3.1
bokeh>=3.3.0
seaborn>=0.13.0
matplotlib>=3.8.0

# 性能优化
numba>=0.58.0  # JIT编译
cupy>=12.0.0  # CUDA加速
taichi>=1.7.0  # 高性能渲染

# 监控告警
slack-sdk>=3.26.0  # Slack通知
discord.py>=2.3.0  # Discord通知
sendgrid>=6.10.0  # 邮件通知
twilio>=8.10.0  # SMS通知

# 文档生成
sphinx>=7.1.0
mkdocs>=1.5.0
pdoc3>=0.10.0
